Show formatted file path in file not found inspection quickfix name
